+++ CHANGELOG.md
@@ -45,6 +45,7 @@ it in future.
 * feat(digi): Use STL vectors for SBT digitisation
 * Change max x of stereo hits to match straw length
 * Get rid of straw diameter dependency in scale factor
+* Assume all tracks are muons during track fit (avoid using MC truth)
 
 ### Removed